# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/v1/quotation-items": {
            "post": {
                "tags": [
                    "Quotations items"
                ],
                "summary": "Create a quotation item",
                "operationId": "6c1d46adb645cb5155ebfc18e6852482",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quotation_id",
                                    "purchase_request_item_id",
                                    "quantity",
                                    "unit_price"
                                ],
                                "properties": {
                                    "quotation_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "purchase_request_item_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 450
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vendor registered successfully"
                                        },
                                        "category": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "IT Equipment"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Books and resources covering software development, computers, programming, and emerging technologies."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/quotation-items/{quotationItem}": {
            "put": {
                "tags": [
                    "Quotations items"
                ],
                "summary": "Update a quotation item",
                "operationId": "70034bba70d39b155c53814a8884578d",
                "parameters": [
                    {
                        "name": "quotationItem",
                        "in": "path",
                        "description": "Quotation item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quotation item updated successfully"
                    },
                    "400": {
                        "description": "Invalid quotation item request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/rfq-vendors": {
            "post": {
                "tags": [
                    "RFQ Vendors"
                ],
                "summary": "Add a vendor to a draft RFQ",
                "operationId": "3435f6ccade69e6712315501a70606f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rfq_id",
                                    "vendor_id"
                                ],
                                "properties": {
                                    "rfq_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vendor_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendor added to RFQ successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "409": {
                        "description": "Vendor already added to this RFQ"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/rfq-vendors/{rfqVendor}": {
            "delete": {
                "tags": [
                    "RFQ Vendors"
                ],
                "summary": "Remove a vendor from a draft RFQ",
                "operationId": "40850915ab433a4c6728c507c384b056",
                "parameters": [
                    {
                        "name": "rfqVendor",
                        "in": "path",
                        "description": "RFQ vendor assignment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendor removed from RFQ successfully"
                    },
                    "400": {
                        "description": "Vendors can only be removed from a draft RFQ"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "RFQ vendor assignment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}